Old classic, remove Integer vs int

+++ base/core/src/main/java/org/openscience/cdk/graph/PathTools.java
@@ -476,7 +476,7 @@ public class PathTools {
                     index = tmp;
                 }
             }
-            qList.remove(index);
+            qList.remove((Integer)index);
             if (index == endNumber) break;
 
             // relaxation
